CORE-252 CORE-254
Fixes for queueing.  Transaction timeout fixed. Fixed status and
exceptions related to timoeout.

+++ edu.harvard.i2b2.crc/src/server/edu/harvard/i2b2/crc/delegate/setfinder/DeleteQueryMasterHandler.java
@@ -69,6 +69,8 @@ public class DeleteQueryMasterHandler extends RequestHandler {
  //           QueryInfoLocal queryInfoLocal = queryInfoLocalHome.create();
             String userId = masterDeleteRequestType.getUserId();
             String masterId = masterDeleteRequestType.getQueryMasterId();
+            if (masterId==null)
+            	throw new I2B2Exception("Null master id sent to DeleteQueryMasterHandler");
             QueryInfoBean query = new QueryInfoBean();
             masterResponseType = query.deleteQueryMaster(getDataSourceLookup(),userId,masterId);
